+++ odata-codegen-runtime/src/main/java/io/github/akbarhusain/odata/runtime/query/NumberProperty.java
@@ -1,6 +1,6 @@
 package io.github.akbarhusain.odata.runtime.query;
 
-public final class NumberProperty<E, N extends Number> extends NumberExpression<N, E> implements PropertyExpression<N> {
+public final class NumberProperty<E, N extends Number> extends NumberExpression<N, E> implements PropertyExpression<E, N> {
 
     public NumberProperty(String edmName, Class<E> entityType) {
         super(edmName, entityType);