+++ odata-codegen-runtime/src/main/java/io/github/akbarhusain/odata/runtime/query/OrderExpression.java
@@ -1,9 +1,9 @@
 package io.github.akbarhusain.odata.runtime.query;
 
-public interface OrderExpression<T> extends Expression<T> {
-    OrderExpression<T> asc();
-    OrderExpression<T> desc();
-    OrderExpression<T> nullsFirst();
-    OrderExpression<T> nullsLast();
+public interface OrderExpression<E, T> extends Expression<T> {
+    OrderExpression<E, T> asc();
+    OrderExpression<E, T> desc();
+    OrderExpression<E, T> nullsFirst();
+    OrderExpression<E, T> nullsLast();
     String getODataPath();
 }