+++ odata-codegen-runtime/src/main/java/io/github/akbarhusain/odata/runtime/query/PropertyExpression.java
@@ -1,5 +1,5 @@
 package io.github.akbarhusain.odata.runtime.query;
 
-public interface PropertyExpression<T> extends OrderExpression<T> {
+public interface PropertyExpression<E, T> extends OrderExpression<E, T> {
     String getEdmName();
 }